added computeAverageForMealsInCafeteriasTags

+++ server/model/cafeteria_rating_models/mealRatingTagsAverage.go
@@ -17,8 +17,8 @@ type MealRatingTagsAverage struct {
 	TagID       int32   `gorm:"column:tagID;foreignKey:tagID;type:int;" json:"tagID"`
 	MealID      int32   `gorm:"column:mealID;foreignKey:mealID;type:int;" json:"mealID"`
 	Average     float32 `gorm:"column:average;type:float;" json:"average" `
-	Min         int     `gorm:"column:min;type:int;" json:"min"`
-	Max         int     `gorm:"column:max;type:int;" json:"max"`
+	Min         int8    `gorm:"column:min;type:int;" json:"min"`
+	Max         int8    `gorm:"column:max;type:int;" json:"max"`
 }
 
 // TableName sets the insert table name for this struct type
